Bugfix from Dave Dodge: tcc doesn't properly handle the following construct:
  typedef [...] a;
  typedef [...] b;

  [...] foo([...])
  {
    a b; /* not parsed correctly */
  }

The problem is that it parses both "a" and "b" as typedef names, and
does nothing as a result.  Although the C99 grammar allows this
parsing, there is a constraint in the narrative that invalidates it
and requires "b" to be treated as a declarator.  This patch causes tcc
to stop looking for typedef names in a declaration specifier list
after it finds the first one; the end result should be that "b"
becomes a local variable and hides the "b" typedef while in scope.