From Dave Dodge:  When a long long value is used as test expression, for example as the controlling expression for an "if" or "while" statement, the tcc i386 backend does not actually test the upper 32 bits.
The problem is that a statement such as "if(x)" ends up in the gtst()
function in i386-gen.c, which fetches the x value and then emits a
test instruction that assumes it's a 32-bit int.  A simple, but
probably not optimal, fix is to treat a long long test the same way as
a floating point test, by inserting an explicit comparison != 0 to
produce a testable 32-bit integer.