Dave Dodge's tcc-0.9.23-pp-number-multi-pasting.patch: There's a bug in the macro_twosharps() function, where a pp-number is not created properly when multiple token pastings are involved.  The sample code below from jun.lambda@gmail.com demonstrates the problem; FLAGS(0) should evaluate to 0 but ends up evaluating to 1:
#define SQUISH(n) (!!((n) & 0xF))
#define HEX2BIN(n) (SQUISH(n))
#define FLAGS(f) HEX2BIN(0x##f##U)
int main () { return FLAGS(0); }

What happens internally is that the first pasting of "0x"+"0" creates
the pp-number "0x0".  The buffer containing that result is then used
directly as the current token string.  When the second pasting begins,
the buffer is reset, destroying the current token string before the
pasting operation.  So instead of pasting "0x0"+"U", it ends up with
""+"U".

A solution is to use a different buffer to hold the current token
contents, so that concatenation won't damage it.  The global tokcstr
is normally used alongside tok and tokc when parsing, so I just used
that.