Pass structures through ?:, test for function pointers too (grischka-2005-09-25 case_1)
This patch was originally posted in grischka's 2005-09-25 email as
required fix case_1 to compile gcc 2.95.  This adds tests to ensure
that both structures and function pointers can pass through the conditional
operator ?:. The tests were tweaked by David A. Wheeler to merge the test cases
into the standard tcc test suite.  This patch also modifies tcc so that
structures can be passed through conditional operators.
Note that this patch does NOT modify tcc to permit function pointers to pass
through ?:, because although such a patch WAS in grischka's patch, an
equivalent change was already in tcc by the time this version of the patch
was created.