Fix signedness of LL shift operators in libtcc1.c (grischka-2005-09-25 case_10) This patch fixes a subtle bug in some long long shift operators in libtcc1.c, caused because the long long shift operators implemented in libtcc1.c didn't cast all their intermediate values to (unsigned) when they should have. This bug caused casting double to unsigned long long (ull) to fail in certain cases, since its implementation used a shift.
This patch modifies portions of libtcc1.c that are only used when libtcc1.c
is compiled with __TINYC__ defined.  (I.E. when tcc compiles itself.)

This patch was originally posted in grischka's 2005-09-25 email as
required fix case_10 to compile gcc 2.95, which modified libtcc1.c's __shldi3
as follows:
-        u.s.high = ((unsigned)u.s.high << b) | (u.s.low >> (32 - b));
+        u.s.high = ((unsigned)u.s.high << b) | ((unsigned)u.s.low >> (32 - b));

David A. Wheeler examined the patch, and noticed that the same error
happened in the other long long shift operators, so he made the same
fix to the rest of the long long shift operators where appropriate.
Notice that in a few cases the arithmetic shift operator does NOT have the
(unsigned) added, because arithmetic right shift is NOT the same as
logical right shift for signed values and it MUST use SIGNED values in
certain cases.

Thankfully, it's REALLY easy to see by visual inspection that this is
what should have been there in the first place. The values
u.s.low and u.s.high are of type "int", which is obviously signed.
The low-level implementation of logical shift operators
should NEVER be shifting signed values, because they only work
as intended with unsigned values. So clearly all of their values should
be cast to unsigned before shifting.  In the arithmetic shift case,
the (unsigned) modification is needed to make sure
that the high bit of the LOWER part of a long long is ignored
(otherwise, it'd be treated as the high bit of the WHOLE word).
Which is all this patch does.  Note that casts like (unsigned) have
a higher precedence than shift, so there's no need to
add parenthesis like this: ((unsigned) u.s.low).
So this patch doesn't add unnecessary parentheses, which would just
make it harder to read.