Fix tcc exit code when unable to create output file.
Michael Somos writes:
I am using the latest CVS version 0.9.24 and found that the exit
status is not getting set on error occurring in the compile and link.

The problem comes from the function "tcc_output_file" in "tccelf.c"
which returns an int which is -1 if an error exists. When it's called
in "tcc.c" the return value is ignored. That code could be improved to

        ret = tcc_output_file(s, outfile) ? 1 : 0;