Workaround for a bug in glibc where #include <regex.h> misuses restrict.
tcc -E boils down to:
  extern int regexec ( const regex_t * __preg ,
  const char * __string , size_t __nmatch ,
  regmatch_t __pmatch [ restrict ] ,
  int __eflags ) ;

and "type name[restrict]" is not a valid variable declaration.  Feed that
to gcc, it'll barf in the same way.  (glibc headers have specific #ifdef
checks for various gcc versions and don't wind up outputting that for them,
but for an unrecognized c99 compiler they output something invalid.)